refactor a lil
