+++ src/rovr/core/preview_container.py
@@ -7,6 +7,7 @@ from contextvars import ContextVar
 from dataclasses import dataclass
 from functools import partial
 from io import BytesIO
+from itertools import batched
 from os import path
 from tempfile import SpooledTemporaryFile
 from time import monotonic, time
@@ -16,19 +17,15 @@ import textual_image.renderable
 import textual_image.widget
 from PIL import Image, UnidentifiedImageError
 from PIL.Image import Image as PILImage
-from rich.cells import cell_len
 from rich.errors import MarkupError
-from rich.syntax import Syntax
 from rich.text import Text
 from textual import events, on, work
 from textual.app import ComposeResult
 from textual.containers import Container
 from textual.css.query import NoMatches
 from textual.dom import DOMNode
-from textual.geometry import Region, Size
+from textual.geometry import Region
 from textual.highlight import guess_language
-from textual.scroll_view import ScrollView
-from textual.strip import Strip
 from textual.timer import Timer
 from textual.widgets import Static
 from textual.widgets.selection_list import Selection
@@ -40,6 +37,7 @@ from rovr.classes.textual_options import (
     FileListSelectionWidget,
 )
 from rovr.components import iterm2_image
+from rovr.components.text_preview import WindowedTextPreview, _decode_text_preview
 from rovr.core import FileList
 from rovr.functions import icons as icon_utils
 from rovr.functions import path as path_utils
@@ -64,16 +62,6 @@ titles = PreviewContainerTitles()
 
 PREVIEWER_GROUP = "previewers"
 TEXT_PREVIEW_CACHE_VERSION = "windowed-v2"
-TEXT_ENCODINGS = (
-    "utf8",
-    "utf16",
-    "utf32",
-    "latin1",
-    "iso8859-1",
-    "mbcs",
-    "ascii",
-    "us-ascii",
-)
 T = TypeVar("T")
 preview_token: ContextVar[object] = ContextVar("preview_token")
 IMAGE_CACHE_SIGNATURE = (
@@ -83,124 +71,6 @@ IMAGE_CACHE_SIGNATURE = (
 
 
 class ExitNow(RuntimeError): ...
-
-
-def _decode_text_preview(data: bytes, truncated: bool) -> tuple[str, int] | None:
-    for encoding in TEXT_ENCODINGS:
-        try:
-            return data.decode(encoding), 0
-        except UnicodeDecodeError as exc:
-            if truncated and exc.end == len(data):
-                try:
-                    return data[: exc.start].decode(encoding), len(data) - exc.start
-                except UnicodeDecodeError:
-                    pass
-    return None
-
-
-class WindowedTextPreview(ScrollView):
-    """Render only the visible portion of a text preview."""
-
-    DEFAULT_CSS = """
-    WindowedTextPreview {
-        width: 1fr;
-        height: 1fr;
-        padding: 0;
-        scrollbar-size: 1 1;
-    }
-    """
-
-    def __init__(
-        self,
-        lines: list[str] | list[Text],
-        *,
-        language: str | None = None,
-        line_numbers: bool = False,
-        classes: str | None = None,
-    ) -> None:
-        super().__init__(classes=classes, can_focus=True)
-        self._lines: list[str] | list[Text] = []
-        self._language = language
-        self._line_numbers = line_numbers
-        self._gutter_width = 0
-        self._rendered_window: tuple[int, int, int, int, list[Strip]] | None = None
-        self.update_preview(lines, language=language, line_numbers=line_numbers)
-
-    def update_preview(
-        self,
-        lines: list[str] | list[Text],
-        *,
-        language: str | None = None,
-        line_numbers: bool = False,
-    ) -> None:
-        self._lines = lines or [""]
-        self._language = language
-        self._line_numbers = line_numbers
-        self._gutter_width = len(str(len(self._lines))) + 3 if line_numbers else 0
-        width = max(
-            cell_len((line.plain if isinstance(line, Text) else line).expandtabs(4))
-            for line in self._lines[:256]
-        )
-        self.virtual_size = Size(width + self._gutter_width, len(self._lines))
-        self._rendered_window = None
-        self.scroll_home(animate=False)
-        self.refresh(layout=True)
-
-    def _render_window(self, start: int, width: int, x: int) -> list[Strip]:
-        end = min(
-            start + max(self.scrollable_content_region.height, 1), len(self._lines)
-        )
-        selected = self._lines[start:end]
-        window_width = max(
-            cell_len((line.plain if isinstance(line, Text) else line).expandtabs(4))
-            for line in selected
-        )
-        if window_width + self._gutter_width > self.virtual_size.width:
-            self.virtual_size = Size(
-                window_width + self._gutter_width, self.virtual_size.height
-            )
-            self._scroll_update(self.virtual_size)
-        if self._language is not None:
-            renderable: Text | Syntax = Syntax(
-                "\n".join(cast(list[str], selected)),
-                lexer=self._language,
-                line_numbers=self._line_numbers,
-                start_line=start + 1,
-                word_wrap=False,
-                tab_size=4,
-                theme=config["theme"]["preview"],
-                background_color=(
-                    "default" if config["theme"]["transparent"] else None
-                ),
-                padding=0,
-            )
-        else:
-            renderable = Text("\n", no_wrap=True, overflow="crop").join(
-                cast(list[Text], selected)
-            )
-
-        options = self.app.console.options.update(width=max(x + width, 1))
-        background = self.visual_style.rich_style
-        return [
-            Strip(segments).crop(x, x + width).adjust_cell_length(width, background)
-            for segments in self.app.console.render_lines(
-                renderable, options, pad=False, new_lines=False
-            )
-        ]
-
-    def render_line(self, y: int) -> Strip:
-        width = self.scrollable_content_region.width
-        height = self.scrollable_content_region.height
-        start = int(self.scroll_offset.y)
-        x = int(self.scroll_offset.x)
-        cache = self._rendered_window
-        if cache is None or cache[:4] != (start, width, height, x):
-            lines = self._render_window(start, width, x)
-            cache = self._rendered_window = (start, width, height, x, lines)
-        try:
-            return cache[4][y]
-        except IndexError:
-            return Strip.blank(width, self.visual_style.rich_style)
 
 
 def _load_cached_image(
@@ -245,7 +115,23 @@ def _load_cached_text(
     if data is None:
         return None
     try:
-        return Text.from_markup(data.decode(), emoji=False)
+        batch: list[str] = [
+            "\n".join(part) for part in batched(data.decode().splitlines(), 4000)
+        ]
+        if len(batch) == 1:
+            return Text.from_markup(batch[0])
+
+        from concurrent.futures import ProcessPoolExecutor
+
+        with ProcessPoolExecutor((os.cpu_count() or 2) // 2) as executor:
+            results = executor.map(
+                Text.from_markup,
+                ("\n".join(part) for part in batched(data.decode().splitlines(), 4000)),
+            )
+        first = next(results)
+        for result in results:
+            first.append_text(result)
+        return first
     except (UnicodeDecodeError, MarkupError):
         return None
